+++ src/main/java/com/sistema/pedidos/repository/ClienteRepository.java
@@ -9,13 +9,13 @@ import org.springframework.data.jpa.repository.Query;
 
 import javax.transaction.Transactional;
 
-public interface ClienteRepository extends JpaRepository<ClientesEntity, Integer>{
+public interface ClienteRepository extends JpaRepository<ClientesEntity, Long>{
     @Query(value = "select c from ClientesEntity c where c.estado = true and c.nombre like ?1% ")
     Page<ClientesEntity> listarClientesPorNombre(String CT_Nombre, Pageable pageable);
     @Transactional
     @Modifying
     @Query(value = "update ClientesEntity c  set c.estado = false where c.idCliente = ?1")
-    int darBajaCliente(Integer id);
+    int darBajaCliente(Long id);
 
     @Query(value = "select c from ClientesEntity c where c.usuario.id = ?1")
     ClientesEntity buscarClientePorIdUsuario(Long idUsuario);
