+++ src/main/java/com/sistema/pedidos/service/ClienteService.java
@@ -6,7 +6,7 @@ import com.sistema.pedidos.commons.GenericServiceApi;
 import com.sistema.pedidos.entity.ClientesEntity;
 import org.springframework.http.ResponseEntity;
 
-public interface ClienteService extends GenericServiceApi<ClientesEntity, Integer> {
+public interface ClienteService extends GenericServiceApi<ClientesEntity, Long> {
